+++ include/maths/CBoostedTree.h
@@ -19,7 +19,13 @@
 #include <memory>
 
 namespace ml {
+namespace core {
+class CPackedBitVector;
+}
 namespace maths {
+class CDataFrameCategoryEncoder;
+class CEncodedDataFrameRowRef;
+
 namespace boosted_tree_detail {
 class MATHS_EXPORT CArgMinLossImpl {
 public:
@@ -122,6 +128,99 @@ public:
 
 class CBoostedTreeImpl;
 
+//! \brief A node of a regression tree.
+//!
+//! DESCRIPTION:\n
+//! This defines a tree structure on a vector of nodes (maintaining the parent
+//! child relationships as indexes into the vector). It holds the (binary)
+//! splitting criterion (feature and value) and the tree's prediction at each
+//! leaf. The intervals are open above so the left node contains feature vectors
+//! for which the feature value is _strictly_ less than the split value.
+//!
+//! During training row masks are maintained for each node (so the data can be
+//! efficiently traversed). This supports extracting the left and right child
+//! node bit masks from the node's bit mask.
+class MATHS_EXPORT CBoostedTreeNode final {
+public:
+    using TSizeSizePr = std::pair<std::size_t, std::size_t>;
+    using TPackedBitVectorPackedBitVectorBoolTr =
+        std::tuple<core::CPackedBitVector, core::CPackedBitVector, bool>;
+    using TNodeVec = std::vector<CBoostedTreeNode>;
+
+public:
+    //! See core::CMemory.
+    static bool dynamicSizeAlwaysZero() { return true; }
+
+    //! Check if this is a leaf node.
+    bool isLeaf() const { return m_LeftChild < 0; }
+
+    //! Get the leaf index for \p row.
+    std::size_t leafIndex(const CEncodedDataFrameRowRef& row,
+                          const TNodeVec& tree,
+                          std::int32_t index = 0) const;
+
+    //! Get the value predicted by \p tree for the feature vector \p row.
+    double value(const CEncodedDataFrameRowRef& row, const TNodeVec& tree) const {
+        return tree[this->leafIndex(row, tree)].m_NodeValue;
+    }
+
+    //! Get the value of this node.
+    double value() const { return m_NodeValue; }
+
+    //! Set the node value to \p value.
+    void value(double value) { m_NodeValue = value; }
+
+    //! Get the gain of the split.
+    double gain() const { return m_Gain; }
+
+    //! Get the total curvature at the rows below this node.
+    double curvature() const { return m_Curvature; }
+
+    //! Get the index of the left child node.
+    std::int32_t leftChildIndex() const { return m_LeftChild; }
+
+    //! Get the index of the right child node.
+    std::int32_t rightChildIndex() const { return m_RightChild; }
+
+    //! Split this node and add its child nodes to \p tree.
+    TSizeSizePr split(std::size_t splitFeature,
+                      double splitValue,
+                      bool assignMissingToLeft,
+                      double gain,
+                      double curvature,
+                      TNodeVec& tree);
+
+    //! Get the row masks of the left and right children of this node.
+    TPackedBitVectorPackedBitVectorBoolTr
+    childrenRowMasks(std::size_t numberThreads,
+                     const core::CDataFrame& frame,
+                     const CDataFrameCategoryEncoder& encoder,
+                     core::CPackedBitVector rowMask) const;
+
+    //! Persist by passing information to \p inserter.
+    void acceptPersistInserter(core::CStatePersistInserter& inserter) const;
+
+    //! Populate the object from serialized data.
+    bool acceptRestoreTraverser(core::CStateRestoreTraverser& traverser);
+
+    //! Get a human readable description of this tree.
+    std::string print(const TNodeVec& tree) const;
+
+private:
+    std::ostringstream&
+    doPrint(std::string pad, const TNodeVec& tree, std::ostringstream& result) const;
+
+private:
+    std::size_t m_SplitFeature = 0;
+    double m_SplitValue = 0.0;
+    bool m_AssignMissingToLeft = true;
+    std::int32_t m_LeftChild = -1;
+    std::int32_t m_RightChild = -1;
+    double m_NodeValue = 0.0;
+    double m_Gain = 0.0;
+    double m_Curvature = 0.0;
+};
+
 //! \brief A boosted regression tree model.
 //!
 //! DESCRIPTION:\n
@@ -152,6 +251,8 @@ public:
     using TRowRef = core::CDataFrame::TRowRef;
     using TLossFunctionUPtr = std::unique_ptr<boosted_tree::CLoss>;
     using TDataFramePtr = core::CDataFrame*;
+    using TNodeVec = std::vector<CBoostedTreeNode>;
+    using TNodeVecVec = std::vector<TNodeVec>;
 
 public:
     ~CBoostedTree() override;
@@ -178,11 +279,11 @@ public:
     //! Get the column containing the model's prediction for the dependent variable.
     std::size_t columnHoldingPrediction(std::size_t numberColumns) const override;
 
+    //! Get the model produced by training if it has been run.
+    const TNodeVecVec& trainedModel() const;
+
     //! Persist by passing information to \p inserter.
     void acceptPersistInserter(core::CStatePersistInserter& inserter) const;
-
-    //! Populate the object from serialized data.
-    bool acceptRestoreTraverser(core::CStateRestoreTraverser& traverser);
 
 private:
     using TImplUPtr = std::unique_ptr<CBoostedTreeImpl>;
