+++ include/maths/CBoostedTreeFactory.h
@@ -67,7 +67,7 @@ public:
     //! Set the target maximum tree depth.
     CBoostedTreeFactory& maxTreeDepth(double maxTreeDepth);
     //! Set the fractional relative tolerance in the target maximum tree depth.
-    CBoostedTreeFactory& maxTreeDepthTolarance(double maxTreeDepthTolarance);
+    CBoostedTreeFactory& maxTreeDepthTolerance(double maxTreeDepthTolerance);
     //! Set the amount we'll shrink the weights on each each iteration.
     CBoostedTreeFactory& eta(double eta);
     //! Set the maximum number of trees in the ensemble.
