+++ include/maths/CBoostedTreeImpl.h
@@ -47,13 +47,15 @@ inline std::size_t predictionColumn(std::size_t numberColumns) {
 //! \brief Implementation of CBoostedTree.
 class MATHS_EXPORT CBoostedTreeImpl final {
 public:
+    using TDoubleVec = std::vector<double>;
     using TMeanAccumulator = CBasicStatistics::SSampleMean<double>::TAccumulator;
     using TMeanVarAccumulator = CBasicStatistics::SSampleMeanVar<double>::TAccumulator;
     using TBayesinOptimizationUPtr = std::unique_ptr<maths::CBayesianOptimisation>;
+    using TNodeVec = CBoostedTree::TNodeVec;
+    using TNodeVecVec = CBoostedTree::TNodeVecVec;
     using TProgressCallback = CBoostedTree::TProgressCallback;
     using TMemoryUsageCallback = CBoostedTree::TMemoryUsageCallback;
     using TTrainingStateCallback = CBoostedTree::TTrainingStateCallback;
-    using TDoubleVec = std::vector<double>;
 
 public:
     static const double MINIMUM_RELATIVE_GAIN_PER_SPLIT;
@@ -83,11 +85,17 @@ public:
     //! Get the feature sample probabilities.
     const TDoubleVec& featureWeights() const;
 
+    //! Get the model produced by training if it has been run.
+    const TNodeVecVec& trainedModel() const;
+
     //! Get the column containing the dependent variable.
     std::size_t columnHoldingDependentVariable() const;
 
     //! Get the number of columns training the model will add to the data frame.
     static std::size_t numberExtraColumnsForTrain();
+
+    //! Get the memory used by this object.
+    std::size_t memoryUsage() const;
 
     //! Estimate the maximum booking memory that training the boosted tree on a data
     //! frame with \p numberRows row and \p numberColumns columns will use.
@@ -98,9 +106,6 @@ public:
 
     //! Populate the object from serialized data.
     bool acceptRestoreTraverser(core::CStateRestoreTraverser& traverser);
-
-    //! Get the memory used by this object.
-    std::size_t memoryUsage() const;
 
 private:
     using TSizeDoublePr = std::pair<std::size_t, double>;
@@ -115,10 +120,6 @@ private:
     using TPackedBitVectorVec = std::vector<core::CPackedBitVector>;
     using TDataFrameCategoryEncoderUPtr = std::unique_ptr<CDataFrameCategoryEncoder>;
     using TDataTypeVec = CDataFrameUtils::TDataTypeVec;
-
-    class CNode;
-    using TNodeVec = std::vector<CNode>;
-    using TNodeVecVec = std::vector<TNodeVec>;
 
     //! \brief Holds the parameters associated with the different types of regularizer
     //! terms available.
@@ -234,175 +235,6 @@ private:
 
         //! Populate the object from serialized data.
         bool acceptRestoreTraverser(core::CStateRestoreTraverser& traverser);
-    };
-
-    //! \brief A node of a regression tree.
-    //!
-    //! DESCRIPTION:\n
-    //! This defines a tree structure on a vector of nodes (maintaining the parent
-    //! child relationships as indexes into the vector). It holds the (binary)
-    //! splitting criterion (feature and value) and the tree's prediction at each
-    //! leaf. The intervals are open above so the left node contains feature vectors
-    //! for which the feature value is _strictly_ less than the split value.
-    //!
-    //! During training row masks are maintained for each node (so the data can be
-    //! efficiently traversed). This supports extracting the left and right child
-    //! node bit masks from the node's bit mask.
-    class CNode final {
-    public:
-        //! See core::CMemory.
-        static bool dynamicSizeAlwaysZero() { return true; }
-
-        //! Check if this is a leaf node.
-        bool isLeaf() const { return m_LeftChild < 0; }
-
-        //! Get the leaf index for \p row.
-        std::size_t leafIndex(const CEncodedDataFrameRowRef& row,
-                              const TNodeVec& tree,
-                              std::int32_t index = 0) const {
-            if (this->isLeaf()) {
-                return index;
-            }
-            double value{row[m_SplitFeature]};
-            bool missing{CDataFrameUtils::isMissing(value)};
-            return (missing && m_AssignMissingToLeft) ||
-                           (missing == false && value < m_SplitValue)
-                       ? tree[m_LeftChild].leafIndex(row, tree, m_LeftChild)
-                       : tree[m_RightChild].leafIndex(row, tree, m_RightChild);
-        }
-
-        //! Get the value predicted by \p tree for the feature vector \p row.
-        double value(const CEncodedDataFrameRowRef& row, const TNodeVec& tree) const {
-            return tree[this->leafIndex(row, tree)].m_NodeValue;
-        }
-
-        //! Get the value of this node.
-        double value() const { return m_NodeValue; }
-
-        //! Set the node value to \p value.
-        void value(double value) { m_NodeValue = value; }
-
-        //! Get the gain of the split.
-        double gain() const { return m_Gain; }
-
-        //! Get the total curvature at the rows below this node.
-        double curvature() const { return m_Curvature; }
-
-        //! Split this node and add its child nodes to \p tree.
-        std::pair<std::size_t, std::size_t> split(std::size_t splitFeature,
-                                                  double splitValue,
-                                                  bool assignMissingToLeft,
-                                                  double gain,
-                                                  double curvature,
-                                                  TNodeVec& tree) {
-            m_SplitFeature = splitFeature;
-            m_SplitValue = splitValue;
-            m_AssignMissingToLeft = assignMissingToLeft;
-            m_LeftChild = static_cast<std::int32_t>(tree.size());
-            m_RightChild = static_cast<std::int32_t>(tree.size() + 1);
-            m_Gain = gain;
-            m_Curvature = curvature;
-            tree.resize(tree.size() + 2);
-            return {m_LeftChild, m_RightChild};
-        }
-
-        //! Get the row masks of the left and right children of this node.
-        auto rowMasks(std::size_t numberThreads,
-                      const core::CDataFrame& frame,
-                      const CDataFrameCategoryEncoder& encoder,
-                      core::CPackedBitVector rowMask) const {
-
-            LOG_TRACE(<< "Splitting feature '" << m_SplitFeature << "' @ " << m_SplitValue);
-            LOG_TRACE(<< "# rows in node = " << rowMask.manhattan());
-            LOG_TRACE(<< "row mask = " << rowMask);
-
-            auto result = frame.readRows(
-                numberThreads, 0, frame.numberRows(),
-                core::bindRetrievableState(
-                    [&](auto& state, TRowItr beginRows, TRowItr endRows) {
-                        core::CPackedBitVector& leftRowMask{std::get<0>(state)};
-                        std::size_t& leftChildNumberRows{std::get<1>(state)};
-                        std::size_t& rightChildNumberRows{std::get<2>(state)};
-                        for (auto row = beginRows; row != endRows; ++row) {
-                            std::size_t index{row->index()};
-                            double value{encoder.encode(*row)[m_SplitFeature]};
-                            bool missing{CDataFrameUtils::isMissing(value)};
-                            if ((missing && m_AssignMissingToLeft) ||
-                                (missing == false && value < m_SplitValue)) {
-                                leftRowMask.extend(false, index - leftRowMask.size());
-                                leftRowMask.extend(true);
-                                ++leftChildNumberRows;
-                            } else {
-                                ++rightChildNumberRows;
-                            }
-                        }
-                    },
-                    std::make_tuple(core::CPackedBitVector{}, std::size_t{0}, std::size_t{0})),
-                &rowMask);
-            auto& masks = result.first;
-
-            for (auto& mask_ : masks) {
-                auto& mask = std::get<0>(mask_.s_FunctionState);
-                mask.extend(false, rowMask.size() - mask.size());
-            }
-
-            core::CPackedBitVector leftRowMask;
-            std::size_t leftChildNumberRows;
-            std::size_t rightChildNumberRows;
-            std::tie(leftRowMask, leftChildNumberRows, rightChildNumberRows) =
-                std::move(masks[0].s_FunctionState);
-            for (std::size_t i = 1; i < masks.size(); ++i) {
-                leftRowMask |= std::get<0>(masks[i].s_FunctionState);
-                leftChildNumberRows += std::get<1>(masks[i].s_FunctionState);
-                rightChildNumberRows += std::get<2>(masks[i].s_FunctionState);
-            }
-            LOG_TRACE(<< "# rows in left node = " << leftRowMask.manhattan());
-            LOG_TRACE(<< "left row mask = " << leftRowMask);
-
-            core::CPackedBitVector rightRowMask{std::move(rowMask)};
-            rightRowMask ^= leftRowMask;
-            LOG_TRACE(<< "# rows in right node = " << rightRowMask.manhattan());
-            LOG_TRACE(<< "left row mask = " << rightRowMask);
-
-            return std::make_tuple(std::move(leftRowMask), std::move(rightRowMask),
-                                   leftChildNumberRows < rightChildNumberRows);
-        }
-
-        //! Get a human readable description of this tree.
-        std::string print(const TNodeVec& tree) const {
-            std::ostringstream result;
-            return this->doPrint("", tree, result).str();
-        }
-
-        //! Persist by passing information to \p inserter.
-        void acceptPersistInserter(core::CStatePersistInserter& inserter) const;
-
-        //! Populate the object from serialized data.
-        bool acceptRestoreTraverser(core::CStateRestoreTraverser& traverser);
-
-    private:
-        std::ostringstream&
-        doPrint(std::string pad, const TNodeVec& tree, std::ostringstream& result) const {
-            result << "\n" << pad;
-            if (this->isLeaf()) {
-                result << m_NodeValue;
-            } else {
-                result << "split feature '" << m_SplitFeature << "' @ " << m_SplitValue;
-                tree[m_LeftChild].doPrint(pad + "  ", tree, result);
-                tree[m_RightChild].doPrint(pad + "  ", tree, result);
-            }
-            return result;
-        }
-
-    private:
-        std::size_t m_SplitFeature = 0;
-        double m_SplitValue = 0.0;
-        bool m_AssignMissingToLeft = true;
-        std::int32_t m_LeftChild = -1;
-        std::int32_t m_RightChild = -1;
-        double m_NodeValue = 0.0;
-        double m_Gain = 0.0;
-        double m_Curvature = 0.0;
     };
 
     //! \brief Maintains a collection of statistics about a leaf of the regression
@@ -739,7 +571,7 @@ private:
     TSizeVec candidateRegressorFeatures() const;
 
     //! Get the root node of \p tree.
-    static const CNode& root(const TNodeVec& tree);
+    static const CBoostedTreeNode& root(const TNodeVec& tree);
 
     //! Get the forest's prediction for \p row.
     static double predictRow(const CEncodedDataFrameRowRef& row, const TNodeVecVec& forest);
